+++ rustler/src/env.rs
@@ -30,6 +30,10 @@ impl<'a, 'b> PartialEq<Env<'b>> for Env<'a> {
     }
 }
 
+///
+#[derive(Clone, Copy, Debug)]
+pub struct SendError;
+
 impl<'a> Env<'a> {
     /// Create a new Env. For the `_lifetime_marker` argument, pass a
     /// reference to any local variable that has its own lifetime, different
@@ -71,12 +75,15 @@ impl<'a> Env<'a> {
     ///
     /// *   The current thread is *not* managed by the Erlang VM.
     ///
+    /// The result indicates whether the send was successful, see also
+    /// [enif\_send](https://www.erlang.org/doc/man/erl_nif.html#enif_send).
+    ///
     /// # Panics
     ///
     /// Panics if the above rules are broken (by trying to send a message from
     /// an `OwnedEnv` on a thread that's managed by the Erlang VM).
     ///
-    pub fn send(self, pid: &LocalPid, message: Term<'a>) {
+    pub fn send(self, pid: &LocalPid, message: Term<'a>) -> Result<(), SendError> {
         let thread_type = unsafe { rustler_sys::enif_thread_type() };
         let env = if thread_type == rustler_sys::ERL_NIF_THR_UNDEFINED {
             ptr::null_mut()
@@ -93,8 +100,14 @@ impl<'a> Env<'a> {
         };
 
         // Send the message.
-        unsafe {
-            rustler_sys::enif_send(env, pid.as_c_arg(), ptr::null_mut(), message.as_c_arg());
+        let res = unsafe {
+            rustler_sys::enif_send(env, pid.as_c_arg(), ptr::null_mut(), message.as_c_arg())
+        };
+
+        if res == 1 {
+            Ok(())
+        } else {
+            Err(SendError)
         }
     }
 
@@ -164,7 +177,7 @@ impl<'a> Env<'a> {
 ///
 ///     fn send_string_to_pid(data: &str, pid: &LocalPid) {
 ///         let mut msg_env = OwnedEnv::new();
-///         msg_env.send_and_clear(pid, |env| data.encode(env));
+///         let _ = msg_env.send_and_clear(pid, |env| data.encode(env));
 ///     }
 ///
 /// There's no way to run Erlang code in an `OwnedEnv`. It's not a process. It's just a workspace
@@ -198,13 +211,15 @@ impl OwnedEnv {
     /// The environment is cleared as though by calling the `.clear()` method.
     /// To avoid that, use `env.send(pid, term)` instead.
     ///
+    /// The result is the same as what `Env::send` would return.
+    ///
     /// # Panics
     ///
     /// Panics if called from a thread that is managed by the Erlang VM. You
     /// can only use this method on a thread that was created by other
     /// means. (This curious restriction is imposed by the Erlang VM.)
     ///
-    pub fn send_and_clear<F>(&mut self, recipient: &LocalPid, closure: F)
+    pub fn send_and_clear<F>(&mut self, recipient: &LocalPid, closure: F) -> Result<(), SendError>
     where
         F: for<'a> FnOnce(Env<'a>) -> Term<'a>,
     {
@@ -214,11 +229,17 @@ impl OwnedEnv {
 
         let message = self.run(|env| closure(env).as_c_arg());
 
-        unsafe {
-            rustler_sys::enif_send(ptr::null_mut(), recipient.as_c_arg(), *self.env, message);
-        }
+        let res = unsafe {
+            rustler_sys::enif_send(ptr::null_mut(), recipient.as_c_arg(), *self.env, message)
+        };
 
         self.clear();
+
+        if res == 1 {
+            Ok(())
+        } else {
+            Err(SendError)
+        }
     }
 
     /// Free all terms in this environment and clear it for reuse.