+++ rustler/src/thread.rs
@@ -41,7 +41,7 @@ where
 {
     let pid = env.pid();
     S::spawn(move || {
-        OwnedEnv::new().send_and_clear(&pid, |env| {
+        let _ = OwnedEnv::new().send_and_clear(&pid, |env| {
             match panic::catch_unwind(|| thread_fn(env)) {
                 Ok(term) => term,
                 Err(err) => {