+++ rustler_tests/lib/rustler_test.ex
@@ -83,6 +83,7 @@ defmodule RustlerTest do
   def threaded_sleep(_), do: err()
 
   def send_all(_, _), do: err()
+  def send(_, _), do: err()
   def whereis_pid(_), do: err()
   def sublists(_), do: err()